+++ Tuneful/Views/Settings/AboutSettingsView.swift
@@ -29,9 +29,11 @@ struct AboutSettingsView: View {
                     
                     Divider()
                     HStack {
+                        Link("Support", destination: URL(string: "https://ko-fi.com/martinfekete")!)
+                            .buttonStyle(.bordered)
                         Link("GitHub", destination: URL(string: "https://github.com/martinfekete10/Tuneful")!)
                             .buttonStyle(.bordered)
-                        Link("Support", destination: URL(string: "https://ko-fi.com/martinfekete")!)
+                        Link("Website", destination: URL(string: "https://martinfekete.com/Tuneful")!)
                             .buttonStyle(.bordered)
                     }
                 }