+++ Tuneful/Views/Settings/MenuBarAppearanceSettingsView.swift
@@ -10,36 +10,32 @@ import Settings
 
 struct MenuBarAppearanceSettingsView: View {
     
-    @AppStorage("showSongInfo") var showSongInfoAppStorage: Bool = true
-    @AppStorage("showMenuBarIcon") var showMenuBarIconAppStorage: Bool = true
-    @AppStorage("trackInfoLength") var trackInfoLengthAppStorage: Double = 20.0
+    @AppStorage("menuBarItemWidth") var menuBarItemWidthAppStorage: Double = 150
     @AppStorage("statusBarIcon") var statusBarIconAppStorage: StatusBarIcon = .albumArt
     @AppStorage("trackInfoDetails") var trackInfoDetailsAppStorage: StatusBarTrackDetails = .artistAndSong
     @AppStorage("popoverBackground") var popoverBackgroundAppStorage: BackgroundType = .transparent
+    @AppStorage("showStatusBarTrackInfo") var showStatusBarTrackInfoAppStorage: ShowStatusBarTrackInfo = .always
     
     // A bit of a hack, binded AppStorage variable doesn't refresh UI, first we read the app storage this way
     // and @AppStorage variable  is updated whenever the state changes using .onChange()
-    @State var showSongInfo: Bool
-    @State var showMenuBarIcon: Bool
-    @State var trackInfoLength: Double
+    @State var menuBarItemWidth: Double
     @State var statusBarIcon: StatusBarIcon
     @State var trackInfoDetails: StatusBarTrackDetails
     @State var popoverBackground: BackgroundType
+    @State var showStatusBarTrackInfo: ShowStatusBarTrackInfo
     
     init() {
-        @AppStorage("showSongInfo") var showSongInfoAppStorage: Bool = true
-        @AppStorage("showMenuBarIcon") var showMenuBarIconAppStorage: Bool = true
-        @AppStorage("trackInfoLength") var trackInfoLengthAppStorage: Double = 20.0
+        @AppStorage("menuBarItemWidth") var menuBarItemWidthAppStorage: Double = 150
         @AppStorage("statusBarIcon") var statusBarIconAppStorage: StatusBarIcon = .albumArt
         @AppStorage("trackInfoDetails") var trackInfoDetailsAppStorage: StatusBarTrackDetails = .artistAndSong
         @AppStorage("popoverBackground") var popoverBackgroundAppStorage: BackgroundType = .transparent
+        @AppStorage("showStatusBarTrackInfo") var showStatusBarTrackInfoAppStorage: ShowStatusBarTrackInfo = .always
         
-        self.showSongInfo = showSongInfoAppStorage
-        self.showMenuBarIcon = showMenuBarIconAppStorage
-        self.trackInfoLength = trackInfoLengthAppStorage
+        self.menuBarItemWidth = menuBarItemWidthAppStorage
         self.statusBarIcon = statusBarIconAppStorage
         self.trackInfoDetails = trackInfoDetailsAppStorage
         self.popoverBackground = popoverBackgroundAppStorage
+        self.showStatusBarTrackInfo = showStatusBarTrackInfoAppStorage
     }
 
     var body: some View {
@@ -48,7 +44,6 @@ struct MenuBarAppearanceSettingsView: View {
                 
                 Settings.Section(label: {
                     Text("Menu bar icon")
-                        .foregroundStyle(!showMenuBarIcon ? .tertiary : .primary)
                 }) {
                     Picker("", selection: $statusBarIcon) {
                         ForEach(StatusBarIcon.allCases, id: \.self) { value in
@@ -60,7 +55,6 @@ struct MenuBarAppearanceSettingsView: View {
                         self.sendTrackChangedNotification()
                     }
                     .pickerStyle(.segmented)
-                    .disabled(!showMenuBarIcon)
                     .frame(width: 200)
                 }
                 
@@ -79,39 +73,25 @@ struct MenuBarAppearanceSettingsView: View {
                     .frame(width: 200)
                 }
                 
-                Settings.Section(title: "Show song info in menu bar") {
-                    Toggle(isOn: $showSongInfo) {
-                        Text("")
-                    }
-                    .onChange(of: showSongInfo) { show in
-                        self.showSongInfoAppStorage = showSongInfo
-                        self.sendTrackChangedNotification()
-                        
-                        if !show {
-                            self.showMenuBarIcon = true
+                Settings.Section(label: {
+                    Text("Show song info in menu bar")
+                }) {
+                    Picker("", selection: $showStatusBarTrackInfo) {
+                        ForEach(ShowStatusBarTrackInfo.allCases, id: \.self) { value in
+                            Text(value.localizedName).tag(value)
                         }
                     }
-                    .toggleStyle(.switch)
-                }
-                
-                Settings.Section(label: {
-                    Text("Show icon in menu bar")
-                        .foregroundStyle(!showSongInfo ? .tertiary : .primary)
-                }) {
-                    Toggle(isOn: $showMenuBarIcon) {
-                        Text("")
-                    }
-                    .onChange(of: showMenuBarIcon) { _ in
-                        self.showMenuBarIconAppStorage = showMenuBarIcon
+                    .onChange(of: showStatusBarTrackInfo) { newValue in
+                        self.showStatusBarTrackInfoAppStorage = showStatusBarTrackInfo
                         self.sendTrackChangedNotification()
                     }
-                    .toggleStyle(.switch)
-                    .disabled(!showSongInfo)
+                    .pickerStyle(.menu)
+                    .frame(width: 200)
                 }
                 
                 Settings.Section(label: {
                     Text("Song info details")
-                        .foregroundStyle(!showSongInfo ? .tertiary : .primary)
+                        .foregroundStyle(self.showStatusBarTrackInfo == .never ? .tertiary : .primary)
                 }) {
                     Picker("", selection: $trackInfoDetails) {
                         ForEach(StatusBarTrackDetails.allCases, id: \.self) { value in
@@ -124,31 +104,31 @@ struct MenuBarAppearanceSettingsView: View {
                     }
                     .pickerStyle(.menu)
                     .frame(width: 200)
-                    .disabled(!showSongInfo)
+                    .disabled(self.showStatusBarTrackInfo == .never)
                 }
                 
                 Settings.Section(label: {
                     Text("Song info max length")
-                        .foregroundStyle(!showSongInfo ? .tertiary : .primary)
+                        .foregroundStyle(self.showStatusBarTrackInfo == .never ? .tertiary : .primary)
                 }) {
                     VStack(alignment: .center) {
-                        Slider(value: $trackInfoLength, in: 10...50, step: 5) {
+                        Slider(value: $menuBarItemWidth, in: 100...300, step: 25) {
                             Text("")
                         } minimumValueLabel: {
-                            Text("10")
+                            Text("100")
                         } maximumValueLabel: {
-                            Text("50")
+                            Text("300")
                         }
-                        .onChange(of: trackInfoLength) { newValue in
-                            self.trackInfoLengthAppStorage = trackInfoLength
+                        .onChange(of: menuBarItemWidth) { newValue in
+                            self.menuBarItemWidthAppStorage = menuBarItemWidth
                             self.sendTrackChangedNotification()
                             NSHapticFeedbackManager.defaultPerformer.perform(NSHapticFeedbackManager.FeedbackPattern.levelChange, performanceTime: .now)
                         }
                         .frame(width: 200)
-                        .disabled(!showSongInfo)
+                        .disabled(self.showStatusBarTrackInfo == .never)
                         
-                        Text("Number of characters: \(Int(trackInfoLength))")
-                            .foregroundStyle(!showSongInfo ? .tertiary : .secondary)
+                        Text("Width: \(Int(self.menuBarItemWidth)) pixels")
+                            .foregroundStyle(self.showStatusBarTrackInfo == .never ? .tertiary : .primary)
                             .font(.callout)
                     }
                 }