+++ release_notes.md
@@ -1,3 +1,13 @@
+# 1.4.0 - Scrolling menu bar text
+* Menu bar text is now scrolling if it does not fit the max width
+* New option to hide menu bar text when nothing is currently playing
+* New keyboard shortcut to quickly switch between Spotify and Apple Music
+* Ability to switch between Spotify and Apple Music from new menu bar submenu
+* Fixed an issue when incremental volume increase would break the UI on max volume
+* Fixed an issue when progress bar would not update until the mini player window is displayed for the first time
+* Minor settings improvements
+* Minor mini player improvements
+
 # 1.3.0 - New mini player options and volume adjustments
 * New feature to customize the look of mini player - you can now choose from 2 styles
 * You can now hide mini player window by right-clicking and choosing Hide window option